+++ tests/test_example/test_example.py
@@ -1,7 +1,7 @@
 """Tests for hello function."""
 import pytest
 
-from parea.main import hello
+from parea.example import hello
 
 
 @pytest.mark.parametrize(
